+++ src/interactions/wheel.component.jsx
@@ -1,5 +1,6 @@
 import { Component } from 'preact';
 import { registerEvents } from '../utils/register-events';
+import { checkModifiers } from '../utils/check-modifiers';
 
 export class Wheel extends Component {
 	_teardownEvents = null;
@@ -14,7 +15,7 @@ export class Wheel extends Component {
 
 	_registerEvents() {
 		this._teardownEvents = registerEvents(this.props.host, {
-			wheel: [this.props.onWheel]
+			wheel: [this.props.onWheel && this._handleWheel]
 		});
 	}
 
@@ -22,6 +23,11 @@ export class Wheel extends Component {
 		this._teardownEvents && this._teardownEvents();
 		this._teardownEvents = null;
 	}
+
+	_handleWheel = event => {
+		checkModifiers(event, this.props.modifiers)
+			&& this.props.onWheel(event);
+	};
 
 	render() {
 		return null;